Fold the mode filter into the filter bar and split out the inspector (fifth feedback round)
GameFilterBar gets the game-mode filter built in behind showGameType instead
of the children/extra-filters pattern, so other game lists can reuse it. The
side panel is split into a generic GameSidePanel (client/games) and the
replay-specific ReplayInspector, ready for future games-page/match-history
panels. useStableCallback is replaced with React 19.2's useEffectEvent.
