+++ server/public/locales/ru/global.json
@@ -469,6 +469,9 @@
       "label": "Фильтры",
       "mapName": "Название карты",
       "matchup": "Матчап",
+      "mode": "Режим",
+      "modeAny": "Любой режим",
+      "modeOthers": "Другие",
       "playerName": "Имя игрока",
       "ranked": "Рейтинговые",
       "sort": {
@@ -1240,11 +1243,6 @@
       "empty": "Реплеев пока нет",
       "emptyBody": "Реплеи сыгранных и просмотренных игр будут появляться здесь автоматически.",
       "filteredCount": "{{shown}} из {{total}} реплеев",
-      "filters": {
-        "mode": "Режим",
-        "modeAny": "Любой режим",
-        "modeOthers": "Другие"
-      },
       "inspector": {
         "empty": "Выберите реплей, чтобы увидеть подробности",
         "parseError": "Не удалось прочитать этот реплей."
